added DTOs and patch functionality

+++ src/main/java/kea/exercise/studentadmin/dtos/student/StudentRequestDTO.java
@@ -3,7 +3,7 @@ package kea.exercise.studentadmin.dtos.student;
 import java.time.LocalDate;
 
 public record StudentRequestDTO (
-    int id,
+    Long id,
     String firstName,
     String middleName,
     String lastName,
