+++ src/main/java/kea/exercise/studentadmin/models/Student.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String firstName;
     private String middleName;
     private String lastName;
@@ -19,7 +19,7 @@ public class Student {
     private boolean graduated;
     private int schoolYear;
 
-    public Student(int id, String firstName, String middleName, String lastName, LocalDate dateOfBirth, House house, boolean prefect, int enrollmentYear, int graduationYear, boolean graduated, int schoolYear) {
+    public Student(Long id, String firstName, String middleName, String lastName, LocalDate dateOfBirth, House house, boolean prefect, int enrollmentYear, int graduationYear, boolean graduated, int schoolYear) {
         this.id = id;
         this.firstName = firstName;
         this.middleName = middleName;
@@ -36,11 +36,11 @@ public class Student {
     public Student() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
